+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
-All notable changes to the "swifttestfilegen" extension will be documented in this file.
+All notable changes to the "SwiftTestFileGen" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.2.0]
+
+- Added `swiftTestFileGen.gotoTestFile.heuristicFilenamePattern` and `swiftTestFileGen.gotoTestFile.useFilenameHeuristics` configurations that enable a simple file search pattern to be used for finding test files.
 
 ## [0.1.0]